+++ neurocaps/_utils/__init__.py
@@ -1,6 +1,6 @@
-from ._check_kwargs import _check_kwargs
-from ._check_parcel_approach import _check_parcel_approach
-from ._pickle_utils import (_convert_pickle_to_dict, _dicts_to_pickles)
-from ._logger import _logger
-from ._analysis import (_CAPGetter, _cap2statmap, _create_display, _run_kmeans, _save_contents)
-from ._extraction import (_TimeseriesExtractorGetter, _check_confound_names, _extract_timeseries)
+from .check_kwargs import _check_kwargs
+from .check_parcel_approach import _check_parcel_approach
+from .pickle_utils import (_convert_pickle_to_dict, _dicts_to_pickles)
+from .logger import _logger
+from .analysis import (_CAPGetter, _cap2statmap, _create_display, _run_kmeans, _save_contents)
+from .extraction import (_TimeseriesExtractorGetter, _check_confound_names, _extract_timeseries)