+++ neurocaps/_utils/_analysis/__init__.py
@@ -1,4 +1,0 @@
-from ._capgetter import _CAPGetter
-from ._cap2statmap import _cap2statmap
-from ._run_kmeans import _run_kmeans
-from ._heatmap_utils import (_create_display, _save_contents)