+++ neurocaps/_utils/analysis/__init__.py
@@ -1,0 +1,4 @@
+from .capgetter import _CAPGetter
+from .cap2statmap import _cap2statmap
+from .run_kmeans import _run_kmeans
+from .heatmap_utils import (_create_display, _save_contents)